+++ tests/Queries/PeopleQueryWithFilter.php
@@ -29,7 +29,7 @@ class PeopleQueryWithFilter extends AbstractQueryWithFilter
      *
      * @return Builder
      */
-    public function execute()
+    public function execute(): Builder
     {
         // Initialize query
         $query = People::query();