+++ CHANGELOG.md
@@ -37,3 +37,8 @@ All notable changes to `datum` will be documented in this file
 ## 0.7.0 - 2021-02-09
 - cut AbstractStaticQuery as it provides no additional functionality compared to AbstractQuery (and use cases result in inconsistent method params & returns)
 - add abstract execute() method to AbstractQuery that will force child classes to conform
+
+
+## 0.7.1 - 2021-02-09
+- make Sfneal\Queries\Traits\HasKeyParam trait for AbstractQuery extensions that require a $modelKey param
+- add improved type hinting to test Filters & Queries